backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
requests==2.32.3
nltk==3.8.1
beautifulsoup4==4.12.3
certifi>=2023.7.22
python-multipart==0.0.6
email-validator==2.1.0
